requirements.txt
# FastAPI + server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sse-starlette>=2.0.0

# Auth
python-jose[cryptography]>=3.3.0
httpx>=0.27.0

# LLM providers
anthropic>=0.28.0
openai>=1.30.0
google-generativeai>=0.7.0

# GCS
google-cloud-storage>=2.17.0

# GCP Secret Manager
google-cloud-secret-manager>=2.20.0

# Workers
feedparser>=6.0.11
PyMuPDF>=1.24.0
pandas>=2.2.0
beautifulsoup4>=4.12.0
rank-bm25>=0.2.2
pyyaml>=6.0.1
slugify>=0.0.1
python-slugify>=8.0.4

# Code execution sandbox
e2b>=0.17.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
